info window and stuff

+++ TowerEngine/Game/code/Game/Ship.h
@@ -15,7 +15,7 @@ struct ship_module_definition {
 
 	ship_module_slot_type SlotType;
 	string DisplayName;
-	real64 ActivationTimeMS;
+	i64 ActivationTimeMS;
 
 	real64 ActivationRange;
 	color ActivationRangeDisplayColor;
